+++ extensions/cpanel/CHANGELOG.md
@@ -1,10 +1,12 @@
 # cPanel Changelog
 
-## [Files + DB - Enhancements] - {PR_MERGE_DATE}
+## [Enhancements: Files & DB + New Account Command ] - {PR_MERGE_DATE}
 
 - Show or Hide "MIME Type" in `Files` via Preferences
 - View "File Content" for small files in `Files`
 - View "Database Schema" in `Databases`
+- View "Account Statistics and Usage" in `Accounts`
+- Change "cPanel Account Password" in `Accounts`
 
 ## [Enhancements] - 2024-09-11